# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_minisource_feedback_internal_models.Attachment:
    properties:
      id:
        type: string
      mimeType:
        type: string
      name:
        type: string
      size:
        type: integer
      storageId:
        description: ID from storage service
        type: string
      uploadedAt:
        type: string
      url:
        type: string
    type: object
  github_com_minisource_feedback_internal_models.Category:
    properties:
      children:
        description: For response - nested children (not stored in DB)
        items:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.Category'
        type: array
      color:
        type: string
      createdAt:
        description: Timestamps
        type: string
      description:
        type: string
      feedbackCount:
        description: Stats
        type: integer
      icon:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      parentId:
        description: Parent category for hierarchical structure
        type: string
      slug:
        type: string
      sortOrder:
        type: integer
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_minisource_feedback_internal_models.CreateCategoryRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 500
        type: string
      icon:
        type: string
      name:
        maxLength: 100
        type: string
      parentId:
        type: string
      slug:
        type: string
      sortOrder:
        type: integer
    required:
    - name
    type: object
...
  github_com_minisource_feedback_internal_models.Feedback:
    properties:
      approvedAt:
        type: string
      attachments:
        description: Attachments (uploaded via storage service)
        items:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.Attachment'
        type: array
      authorAvatar:
        type: string
      authorEmail:
        type: string
      authorId:
        description: Author information
        type: string
      authorName:
        type: string
      categoryId:
        description: Category and tags
        type: string
      categoryName:
        type: string
      commentCount:
        type: integer
      completedAt:
        type: string
      createdAt:
        description: Timestamps
        type: string
      deletedAt:
        description: Soft delete
        type: string
      deletedBy:
        type: string
      description:
        type: string
      downvotes:
        type: integer
      id:
        type: string
      isAnonymous:
        type: boolean
      officialResponse:
        allOf:
        - $ref: '#/definitions/github_com_minisource_feedback_internal_models.OfficialResponse'
        description: Admin response
      status:
        $ref: '#/definitions/github_com_minisource_feedback_internal_models.FeedbackStatus'
      tags:
        items:
          type: string
        type: array
      tenantId:
        type: string
      title:
        type: string
      trendingScore:
        description: Trending score (calculated based on votes, comments, views, and
          time decay)
        type: number
      updatedAt:
        type: string
      upvotes:
        description: Voting
        type: integer
      viewCount:
        description: Engagement metrics
        type: integer
      voteScore:
        description: upvotes - downvotes
        type: integer
    type: object
...
  github_com_minisource_feedback_internal_models.FeedbackStatus:
    enum:
    - pending
    - approved
    - rejected
    - under_review
    - planned
    - in_progress
    - completed
    - closed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusRejected
    - StatusUnderReview
    - StatusPlanned
    - StatusInProgress
    - StatusCompleted
    - StatusClosed
  github_com_minisource_feedback_internal_models.OfficialResponse:
    properties:
      content:
        type: string
      respondedAt:
        type: string
      respondedBy:
        type: string
      respondedByName:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_minisource_feedback_internal_models.Setting:
    properties:
      createdAt:
        description: Timestamps
        type: string
      description:
        type: string
      id:
        type: string
      isPublic:
        description: If true, visible to non-admins
        type: boolean
      key:
        type: string
      tenantId:
        type: string
      type:
        $ref: '#/definitions/github_com_minisource_feedback_internal_models.SettingType'
      updatedAt:
        type: string
      updatedBy:
        type: string
      value: {}
    type: object
  github_com_minisource_feedback_internal_models.SettingType:
    enum:
    - string
    - number
    - boolean
    - json
    type: string
    x-enum-varnames:
    - SettingTypeString
    - SettingTypeNumber
    - SettingTypeBool
    - SettingTypeJSON
...
  github_com_minisource_feedback_internal_models.UpdateCategoryRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 500
        type: string
      icon:
        type: string
      isActive:
        type: boolean
      name:
        maxLength: 100
        type: string
      parentId:
        type: string
      slug:
        type: string
      sortOrder:
        type: integer
    type: object
...
  github_com_minisource_feedback_internal_models.UpdateSettingRequest:
    properties:
      description:
        type: string
      isPublic:
        type: boolean
      key:
        type: string
      type:
        type: string
      value: {}
    required:
    - key
    - value
    type: object
...
host: localhost:5012
info:
  contact: {}
  description: Feedback management service for Minisource
  title: Feedback Service API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create category request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Category'
      summary: Create category
      tags:
      - Admin - Categories
  /admin/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete category
      tags:
      - Admin - Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update category request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Category'
      summary: Update category
      tags:
      - Admin - Categories
  /admin/feedback/{id}/approve:
    post:
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Feedback'
      summary: Approve feedback
      tags:
      - Admin - Moderation
  /admin/feedback/{id}/reject:
    post:
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Feedback'
      summary: Reject feedback
      tags:
      - Admin - Moderation
  /admin/feedback/{id}/response:
    post:
      consumes:
      - application/json
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: string
      - description: Official response
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Feedback'
      summary: Add official response
      tags:
      - Admin - Moderation
  /admin/feedback/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Feedback'
      summary: Update feedback status
      tags:
      - Admin - Moderation
  /admin/feedback/pending:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get pending feedback
      tags:
      - Admin - Moderation
  /admin/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_minisource_feedback_internal_models.Setting'
            type: array
      summary: Get all settings
      tags:
      - Admin - Settings
    put:
      consumes:
      - application/json
      parameters:
      - description: Update setting request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_feedback_internal_models.UpdateSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update setting
      tags:
      - Admin - Settings
...
  /categories:
    get:
      parameters:
      - description: Only active categories
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_minisource_feedback_internal_models.Category'
            type: array
      summary: List categories
      tags:
      - Categories
  /categories/{id}:
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_feedback_internal_models.Category'
      summary: Get category
      tags:
      - Categories
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"